Create info
In this example, we create a simple animation of a bouncing object represented by the letter 'O'. Each frame of the animation is displayed by printing a line with a different number of spaces before the 'O'. The animation is created by iterating through the frames, clearing the screen, printing the current frame, and pausing for a short duration using std::this_thread::sleep_for. Adjust the value of animationFrames and the sleep duration to modify the animation speed and duration.

Note that this example assumes your terminal supports ANSI escape codes for clearing the screen ("\033[2J\033[1;1H"). If your terminal does not support these codes, you may need to modify the clearScreen() function accordingly.